+++ source/WebLogger/Commands/HelpCommandHandler.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Text;
-using WebLogger.Display;
+using WebLogger.Render;
 
 namespace WebLogger.Commands
 {
