--- source/WebLogger/Display/Direction.cs
@@ -1,11 +1,0 @@
-﻿namespace WebLogger.Display
-{
-    /// <summary>
-    /// Html Opening Direction. 
-    /// </summary>
-    internal enum Direction
-    {
-        Open,
-        Close
-    }
-}
