--- source/WebLogger/Display/HtmlElementExtensions.cs
@@ -1,25 +1,0 @@
-﻿//using System.Text;
-
-//namespace WebLogger.Display
-//{
-//    public static class HtmlElementExtensions
-//    {
-//        public static HtmlElement Append(this HtmlElement element, string value)
-//        {
-//            return element.Append(value);
-//        }
-//        public static HtmlElement Append(this HtmlElement element, HtmlElement value)
-//        {
-//            return element.Append(value);
-//        }
-//        public static HtmlElement Append(this HtmlElement element, StringBuilder value)
-//        {
-//            return element.Append(value);
-//        }
-        
-//        public static HtmlElement Insert(this HtmlElement element, HtmlElement value)
-//        {
-//            return element.Insert(value);
-//        }
-//    }
-//}
