+++ source/WebLogger/Render/StringBuilderMarkupExtensions.cs
@@ -1,7 +1,6 @@
-﻿using System.Reflection;
-using System.Text;
+﻿using System.Text;
 
-namespace WebLogger.Display
+namespace WebLogger.Render
 {
     internal static class StringBuilderMarkupExtensions
     {
